requirements.txt
aiodns==3.0.0
aiofiles==0.6.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.4.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
dacite==1.6.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
pycares==4.2.0
pycparser==2.21
PyNaCl==1.5.0
requests==2.28.0
typing_extensions==4.2.0
update-checker==0.18.0
urllib3==1.26.9
yarl==1.7.2
